Feat : USER Create & Post API implements 1. Feat : Board , Comment 관련 메소드 추가 2. Feat : Response 관련 클래스 및 인터페이스 추가 3. Feat : User 관련 데이터베이스 연동 작업인 POST API & READ API 추가

+++ src/main/java/com/example/jungleboarding/board/BoardRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
-
+public interface BoardRepository extends JpaRepository<Board, Integer> {
+    ;
 }
